Recording Screen Split
Playback now takes place in the PlaybackScreen activity. Creating this
activity requires that the recorded filename (absolute path) be passed
in. In addition, currently a visualization file must be included.

Recording only really changed in the sense that playback functionality
has been removed from this activity.
