+++ RecordingApp/app/src/main/AndroidManifest.xml
@@ -26,6 +26,7 @@
         </activity>
 
         <activity android:name=".Recording.RecordingScreen" android:screenOrientation="landscape" android:noHistory="true"/>
+        <activity android:name=".Playback.PlaybackScreen" android:screenOrientation="landscape" android:noHistory="true"/>
         <activity android:name=".FilesPage.AudioFiles" android:screenOrientation="portrait" />
         <activity android:name=".SettingsPage.Settings" android:screenOrientation="portrait"
             android:windowSoftInputMode="stateUnchanged"/>
