+++ RecordingApp/app/src/main/java/wycliffeassociates/recordingapp/AudioVisualization/WaveformView.java
@@ -56,7 +56,7 @@ public class WaveformView extends CanvasView {
         }
         redraw();
         WavPlayer.checkIfShouldStop();
-        if(!WavPlayer.isPlaying()){
+        if(WavPlayer.exists() && !WavPlayer.isPlaying()){
             manager.enablePlay();
         }
     }
